Added method to check if ROM was changed.
Originally rom_stats detected the flashing of a new ROM by reading the preference ANONYMOUS_FIRST_BOOT. If this preference was missing, a fresh boot after a ROM flash was expected. To achieve this the rom_stats preferences had to be deleted from updater_script during the flashing of the ROM. This was done by rm -rf /data/data/android.romstats.
If the updater_script did not remove rom_stats preferences then rom_stats could not detect that a new ROM was flashed.

This method makes it possible to find a ROM change independent from the action in the actions in the updater_script. The method simple compares the last ROM version saved in the preferences ("0" if no preferences exist) with the ROM version saved in build.prop. If the two strings are different a new ROM version or a complete different ROM was flashed.